+++ jarvis/__init__.py
@@ -2,7 +2,7 @@ import os
 
 import pynotification
 
-version = "7.1.9"
+version = "7.2.0"
 
 install_script = os.path.join(os.path.dirname(__file__), 'lib', 'install.sh')